backend/requirements.txt
pipecat-ai[daily,deepgram,groq,silero]
fastapi
uvicorn[standard]
python-dotenv
aiohttp
aiosqlite
loguru
